+++ common/minimal-modules.nix
@@ -3,6 +3,6 @@ let
   inherit (self.common) options package-sets standardise-nix;
 in
 builtins.mapAttrs (package-set: _: [
-  options.${package-set}
+  options.${package-set}.minimal
   standardise-nix.${package-set}
 ]) package-sets